updated navbar file to hold everything in one file
