+++ src/reducers/appReducer.jsx
@@ -7,6 +7,11 @@ export const appReducer = (state, action) => {
     case "SET_IS_MOBILE":
       // console.log("isMobile", action.payload);
       return { ...state, isMobile: action.payload };
+      break;
+
+    case "SET_LANGUAGE":
+      return { ...state, language: action.payload };
+      break;
     default:
       return state;
   }
